+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Changed
 - Handle data packet trailers if present.
+- Show controllee/controller IDs in command packets.
 - Display Discrete I/O fields in hexadecimal.
 
 ## [0.1.0] - 2019-12-30